intermediate version 3.30 Herman
